+++ packages/shared/src/UI/modals/SelectFungibleTokenModal/index.tsx
